+++ Sanctuary.Census.Builder/CollectionBuilders/ImageCollectionsBuilder.cs
@@ -7,22 +7,24 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using MImage = Sanctuary.Census.Common.Objects.Collections.Image;
 using MImageSet = Sanctuary.Census.Common.Objects.Collections.ImageSet;
+using MImageSetDefault = Sanctuary.Census.Common.Objects.Collections.ImageSetDefault;
 
 namespace Sanctuary.Census.Builder.CollectionBuilders;
 
 /// <summary>
-/// Builds the <see cref="MImageSet"/> collection.
+/// Builds the <see cref="MImage"/> and <see cref="MImageSet"/> collections.
 /// </summary>
-public class ImageSetCollectionBuilder : ICollectionBuilder
+public class ImageCollectionsBuilder : ICollectionBuilder
 {
     private readonly IClientDataCacheService _clientDataCache;
 
     /// <summary>
-    /// Initializes a new instance of the <see cref="ImageSetCollectionBuilder"/> class.
+    /// Initializes a new instance of the <see cref="ImageCollectionsBuilder"/> class.
     /// </summary>
     /// <param name="clientDataCache">The client data cache.</param>
-    public ImageSetCollectionBuilder
+    public ImageCollectionsBuilder
     (
         IClientDataCacheService clientDataCache
     )
@@ -89,6 +91,49 @@ public class ImageSetCollectionBuilder : ICollectionBuilder
             }
         }
 
+        Dictionary<uint, MImage> builtImages = new();
+        foreach (Image image in _clientDataCache.Images)
+        {
+            builtImages.Add(image.ID, new MImage
+            (
+                image.ID,
+                System.IO.Path.GetFileNameWithoutExtension(image.FileName),
+                $"/files/ps2/images/static/{image.ID}.png"
+            ));
+        }
+        await dbContext.UpsertCollectionAsync(builtImages.Values, ct).ConfigureAwait(false);
+
+        Dictionary<uint, MImageSetDefault> builtDefaultSets = new();
+        foreach (MImageSet set in builtSets)
+        {
+            if (!builtDefaultSets.TryGetValue(set.ImageSetID, out MImageSetDefault? existing))
+            {
+                builtDefaultSets.Add(set.ImageSetID, SetToDefault(set));
+                continue;
+            }
+
+            if (existing.TypeID == (uint)ImageSetType.Type.Large)
+                continue;
+
+            if (existing.TypeID > set.TypeID)
+                continue;
+
+            builtDefaultSets[set.ImageSetID] = SetToDefault(set);
+        }
+
+        await dbContext.UpsertCollectionAsync(builtDefaultSets.Values, ct).ConfigureAwait(false);
+        // This must occur after upserting the default sets, as the builder clears the collection
         await dbContext.UpsertCollectionAsync(builtSets, ct).ConfigureAwait(false);
     }
+
+    private static MImageSetDefault SetToDefault(MImageSet set)
+        => new
+        (
+            set.ImageSetID,
+            set.ImageID,
+            set.Description,
+            set.TypeID,
+            set.TypeDescription,
+            set.ImagePath
+        );
 }