+++ Sanctuary.Census.Common/Objects/Collections/SkillSet.cs
@@ -4,6 +4,17 @@ using Sanctuary.Census.Common.Objects.CommonModels;
 
 namespace Sanctuary.Census.Common.Objects.Collections;
 
+/// <summary>
+/// Represents a skill set.
+/// </summary>
+/// <param name="SkillSetID">The ID of the skill set.</param>
+/// <param name="SkillPoints">Unknown.</param>
+/// <param name="RequiredItemID">Unknown.</param>
+/// <param name="Name">The name of the skill set.</param>
+/// <param name="Description">The description of the skill set.</param>
+/// <param name="ImageSetID">The ID of the skill set's image set.</param>
+/// <param name="ImageID">The ID of the set's default image.</param>
+/// <param name="ImagePath">The relative path to the set's default image.</param>
 [Collection]
 public record SkillSet
 (